Added size checks and proper room splitting
I added a check to see if it is even possible to generate a room with the current rooms width or height.
If it's not possible to generate a room then it will skip the generation of the room.
It will now properly split the room. it checks if it reached a certain level and then switches the horizontal splitting to the vertical splitting.