Spitting Dungeon V3
I changed the while loop argument to the depth instead of the amount of rooms.
Each time it reaches a deeper level of splitting, it will change the color and if it will split horizontally or vertically.

Changed the void of splitting the rooms into a coroutine to remove unnecesarry lists.

CanSplit void bool will now check for the next splitting posibility if the first one doesn't work.

Added a room deduction variable, that will deduct the maximum rooms able to be created each level.
The deduction will be added each time the current split isnt possible.