List of final rooms and dictionary added V1
I created a list with all the completed rooms so I can find them easier.
I turned the list of created rooms into a dictionary.
With the dictionary you will be able to sort through the depth layers and find the room sizes you want.
This is used for getting the most recently created rooms.