Door Creation V1 DONE - completed room adjustements - Inspector clean up
Fixed the bug where a door would be placed in the bottom left due to it being placed at a doors width.
Fixed the bug where the door could be placed further then it should be placed.
Made small adjustements to how the completed rooms are added to the list.
This is not a fix but a cleaner version of it.
Inspector now has headers and a logical order for easier use.