Optimized Adding doors & changed door placement.
Optimized the way doors are added. I used my own code to get the overlap/intersection of two RectInts.
But I found out the algorithmUtils has a usefull intersect function that makes find the overlap easier.
Created a dictionary that keeps track which rooms a door is connected to.
Slightly adjusted the bubble sorter for easier use.