Added Node Graph, Room-Door Connections & better door creation
I added a node graph that is toggle able in the inspector
The node graph goes through the 2 dictionaries that connect the doors and rooms

Each time a door is created, the 2 dictionaries will be filled with a door and a room that connects to one another.

Through the dictionaries the check for connected rooms is improved.